refactor: extract code into libs.ts
